refactor(pipelines): extract rendering and spatial config seams
