+++ docs/architecture/MONOLITH_DECOMPOSITION_TARGETS.md
@@ -181,6 +181,8 @@ This table is the canonical progress ledger for extraction PRs that draw from th
 | Target 4A — Rendering config/result types | `pipelines/rendering_4k_pipeline.py` | `pipelines/rendering_4k/types.py` | Landed | This PR: rendering config/result types extracted; ranking unchanged pending governance refresh |
 | Target 4B — Rendering pure image stage functions | `pipelines/rendering_4k_pipeline.py` | `pipelines/rendering_4k/stages.py` | Landed | This PR: rendering pure image stage functions extracted; ranking unchanged pending governance refresh |
 | Target 4C — Rendering quality and memory helpers | `pipelines/rendering_4k_pipeline.py` | `pipelines/rendering_4k/quality.py` | Landed | This PR: rendering quality assessor and GPU memory manager extracted; ranking unchanged pending governance refresh |
+| Target 4D — Rendering pipeline orchestration | `pipelines/rendering_4k_pipeline.py` | `pipelines/rendering_4k/pipeline.py` | Landed | This PR: rendering pipeline orchestration extracted behind legacy CLI/import shim; ranking unchanged pending governance refresh |
+| Target 5A — Spatial AI pipeline config | `spatial_ai/orchestration/pipeline.py` | `spatial_ai/orchestration/config.py` | Landed | This PR: spatial pipeline config model and validation helpers extracted; ranking unchanged pending governance refresh |
 
 ---
