+++ policy/json_raw_approved_modules.txt
@@ -33,7 +33,7 @@ src/transformation_portal/perceptual/baseline.py
 src/transformation_portal/perceptual/tracker.py
 src/transformation_portal/pipeline_unified.py
 src/transformation_portal/pipelines/quality_feedback_bridge.py
-src/transformation_portal/pipelines/rendering_4k_pipeline.py
+src/transformation_portal/pipelines/rendering_4k/pipeline.py
 src/transformation_portal/pipelines/unified_luxury_pipeline.py
 src/transformation_portal/spatial_ai/ingest/linear_decoder.py
 src/transformation_portal/spatial_ai/orchestration/graph/artifact_store.py
