+++ src/transformation_portal/spatial_ai/orchestration/__init__.py
@@ -32,8 +32,9 @@ Architecture (ADR-027, ADR-028):
 
 from __future__ import annotations
 
+from .config import PipelineConfig
 from .error_handler import ErrorHandler, ErrorRecoveryStrategy, PipelineError
-from .pipeline import PipelineConfig, PipelineResult, SpatialAIPipeline
+from .pipeline import PipelineResult, SpatialAIPipeline
 from .progress_tracker import ProgressEvent, ProgressTracker
 from .resource_manager import ResourceLimits, ResourceManager
